requirements.txt
boto3 
s3fs
python-dotenv
fastapi[standard]
fsspec
# deprecated
#gunicorn
uvicorn
openpyxl 
pandas 
#psycopg2-binary
pydantic 
python-multipart
requests>=2.31.0,<3
scikit-learn==1.4.2
typing
#matplotlib
#seaborn
#plotly
#streamlit
mlflow==3.7.0
cloudpickle==3.1.0
colorama==0.4.6
defusedxml==0.7.1
gmpy2==2.2.1
numpy==2.1.3
packaging==25.0
pandas==2.3.3
pyyaml==6.0.3
torch==2.5.0
torchvision==0.20.0
tqdm==4.67.1
typing-extensions==4.15.0
pillow>=10.0.0